refactor(Component): 🚑️ Change "navbar" to "ucsc-header" for WP consistency
