+++ src/components/trss-ucsc-header/test/trss-ucsc-header.e2e.ts
@@ -1,11 +1,11 @@
 import { newE2EPage } from '@stencil/core/testing';
 
-describe('trss-ucsc-navbar', () => {
+describe('trss-ucsc-header', () => {
   it('renders', async () => {
     const page = await newE2EPage();
-    await page.setContent('<trss-ucsc-navbar></trss-ucsc-navbar>');
+    await page.setContent('<trss-ucsc-header></trss-ucsc-header>');
 
-    const element = await page.find('trss-ucsc-navbar');
+    const element = await page.find('trss-ucsc-header');
     expect(element).toHaveClass('hydrated');
   });
 });
